backend/requirements.txt
amqp==5.1.1
ansicon==1.89.0
argcomplete==2.0.0
arrow==1.2.2
asgiref==3.5.1
asttokens==2.0.5
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bleach==3.2.1
blessed==1.19.1
boto3==1.24.47
build==0.8.0
celery==5.2.7
cement==2.8.2
certifi==2020.12.5
cffi==1.15.0
cfn-flip==1.3.0
chardet==3.0.4
charset-normalizer==2.0.12
chromedriver==2.24.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
crayons==0.4.0
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.4
django-background-tasks==1.2.5
django-celery-beat==2.3.0
django-compat==1.0.15
django-cors-headers==3.13.0
django-heroku==0.3.1
django-picklefield==3.0.1
django-q==1.3.9
django-rest-knox==4.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
durationpy==0.5
entrypoints==0.4
executing==0.8.3
fonttools==4.33.3
future==0.16.0
gunicorn==20.1.0
h11==0.13.0
hjson==3.0.2
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jinxed==1.2.0
jmespath==0.10.0
joblib==1.1.0
jupyter-client==7.3.1
jupyter-core==4.10.0
kappa==0.6.0
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
multitasking==0.0.11
nest-asyncio==1.5.5
numpy==1.22.3
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.2.0
pip-tools==6.8.0
placebo==0.10.0
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pymongo==3.12.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.2
pytz==2022.1
PyYAML==5.4.1
pyzmq==23.2.0
redis==3.5.3
requests==2.26.0
s3transfer==0.6.0
scikit-learn==1.1.0
scipy==1.8.0
selenium==4.3.0
semantic-version==2.8.5
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
stack-data==0.2.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
trio==0.21.0
trio-websocket==0.9.2
troposphere==4.1.0
tzdata==2022.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.0.0
wsgi-request-logger==0.4.6
wsproto==1.1.0
XlsxWriter==3.0.3
yfinance==0.1.74
zappa==0.54.2
